% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center_select.R
\name{center_select}
\alias{center_select}
\title{Generate center select}
\usage{
center_select(data, number)
}
\arguments{
\item{data}{training datasets}

\item{number}{number of center}
}
\description{
Generate center select
}
\examples{
\dontrun{
csv1 = read.csv("data.csv", sep = ",", header = TRUE)
cs = center_select(csv1, 666)
}
}
